refactor(remotecontrol): replace remote module for electron 14+
